Anagram
Checks whether a given string contains the same number of characters, except that the sequence of characters can be changed in both strings.

For example, suppose we have two strings: "CAB" and "ABC". Both contain the same characters in these strings, but the arrangements of the first strings' characters are different from the second strings, and the process is termed the anagram.